Move DependencyManager to `dependencies.py`
