+++ coreblocks/params/__init__.py
@@ -4,3 +4,4 @@ from .genparams import *  # noqa: F401
 from .layouts import *  # noqa: F401
 from .fu_params import *  # noqa: F401
 from .keys import *  # noqa: F401
+from .dependencies import *  # noqa: F401
